Financial_analysis/requirements.txt
phidata
python-dotenv
yfinance
duckduckgo-search
fastapi
uvicorn
groq
packaging
openai
gpt4all
langchain
langchain_community